pyproject.toml
[project]
name = "paper-newsletter"
version = "0.1.0"
description = "Daily LLM agent paper newsletter → Slack"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1",
    "httpx>=0.27",
    "google-genai>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
paper-newsletter = "src.main:main"
